requirements.txt
langchain
langchain-community
langchain-core
langchain-chroma
langchain-google-genai
google-generativeai
chromadb
python-dotenv